// happytrip_frontend/src/app/models/generateNewTicketResponseModel.ts
export interface generateNewTicketResponse {
    ticketId: string
    busName: string
    source: string
    destination: string
    boardingPoint: string
    droppingPoint: string
    departureDate: string
    departureTime: string
    coachType: string
    userId: string
    userName: string
    emailId: string
    phoneNumber: string
    ticketStatus: string
    cancellationDate: any
    refundedAmount: any
    busId: string
    seatsBooked: string
    createdDate: string
    totalAmountPaid: string
  }
  